Fix warnings

+++ intern/src/base/base_serialize_std_string.h
@@ -23,8 +23,8 @@ namespace SER
 	template<class TArchive, typename TObject, class TTraits, class TAllocator>
     inline void Write(TArchive& _rArchive, const std::basic_string<TObject, TTraits, TAllocator>& _rString)
     {
-        _rArchive.template BeginCollection<TObject>(_rString.size());
-        _rArchive.WriteCollection(_rString.data(), _rString.size());
+        _rArchive.template BeginCollection<TObject>(static_cast<unsigned int>(_rString.size()));
+        _rArchive.WriteCollection(_rString.data(), static_cast<unsigned int>(_rString.size()));
         _rArchive.template EndCollection<TObject>();
     }
